FEAT: updated faction enlistment with models, fixed org schema
